Provide a meaningful default network interface
The code as it is today, defaults to eth0 for the network interface if one
was not provided via the config file. Added get_default_if to scan system
interfaces to find one that has a MAC address and that is not null. For ex,
loopback devices in macbooks don't have hwaddr, whereas in linux, it does
have hwaddr but its all zeros. Subject to system configuration, it
defaults to returning en0 in macbook pros and eth0 in linux systems.

This is meant to ease development scenarios. It is strongly suggested that
a proper interface is specified in production systems.